added logic to ReporteFrame to show a report to the user

+++ src/main/java/com/alura/jdbc/view/ControlDeStockFrame.java
@@ -256,7 +256,7 @@ public class ControlDeStockFrame extends JFrame {
         var categoria = (Categoria) comboCategoria.getSelectedItem();
 
 
-        this.productoController.guardar(producto, categoria.getId());
+        this.productoController.guardar(producto, (Integer) categoria.getId());
 
         JOptionPane.showMessageDialog(this, "Registrado con éxito!");
